+++ NEWS.md
@@ -1,4 +1,4 @@
-# brms 2.13.0++
+# brms 2.13.3
 
 ### New Features